requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
envparse==0.2.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
pydantic_core==2.14.5
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
